Struct zircon_syscall::Syscall

pub struct Syscall<'a> {
    pub thread: &'a CurrentThread,
    pub thread_fn: ThreadFn,
}

Fields
thread: &'a CurrentThreadthread_fn: ThreadFn
Implementations

impl Syscall<'_>

pub fn sys_channel_read(
&self,
handle_value: HandleValue,
options: u32,
bytes: UserOutPtr<u8>,
handles: usize,
num_bytes: u32,
num_handles: u32,
actual_bytes: UserOutPtr<u32>,
actual_handles: UserOutPtr<u32>,
is_etc: bool
) -> ZxResult

Read/Receive a message from a channel.

pub fn sys_channel_write(
&self,
handle_value: HandleValue,
options: u32,
user_bytes: UserInPtr<u8>,
num_bytes: u32,
user_handles: UserInPtr<HandleValue>,
num_handles: u32
) -> ZxResult

Write a message to a channel.

pub fn sys_channel_create(
&self,
options: u32,
out0: UserOutPtr<HandleValue>,
out1: UserOutPtr<HandleValue>
) -> ZxResult

Create a new channel.

pub async fn sys_channel_call_noretry(
&self,
handle_value: HandleValue,
options: u32,
deadline: Deadline,
user_args: UserInPtr<ChannelCallArgs>,
actual_bytes: UserOutPtr<u32>,
actual_handles: UserOutPtr<u32>
) -> ZxResult

pub fn sys_channel_call_finish(
&self,
deadline: Deadline,
user_args: UserInPtr<ChannelCallArgs>,
_actual_bytes: UserOutPtr<u32>,
_actual_handles: UserOutPtr<u32>
) -> ZxResult

pub fn sys_channel_write_etc(
&self,
handle: HandleValue,
options: u32,
user_bytes: UserInPtr<u8>,
num_bytes: u32,
user_handles: UserInOutPtr<HandleDisposition>,
num_handles: u32
) -> ZxResult

Write a message to a channel.

impl Syscall<'_>

pub fn sys_cprng_draw_once(&self, buf: UserOutPtr<u8>, len: usize) -> ZxResult

Draw random bytes from the kernel CPRNG.

This data should be suitable for cryptographic applications.

Clients that require a large volume of randomness should consider using these bytes to seed a user-space random number generator for better performance.

impl Syscall<'_>

pub fn sys_iommu_create(
&self,
resource: HandleValue,
type_: u32,
desc: UserInPtr<u8>,
desc_size: usize,
out: UserOutPtr<HandleValue>
) -> ZxResult

Create a new object in the kernel representing an IOMMU device.

pub fn sys_bti_create(
&self,
iommu: HandleValue,
options: u32,
bti_id: u64,
out: UserOutPtr<HandleValue>
) -> ZxResult

Creates a new bus transaction initiator.

iommu: HandleValue, a handle to an IOMMU. options: u32, must be 0 (reserved for future definition of creation flags). bti_id: u64, a hardware transaction identifier for a device downstream of that IOMMU.

pub fn sys_bti_pin(
&self,
bti: HandleValue,
options: u32,
vmo: HandleValue,
offset: usize,
size: usize,
addrs: UserOutPtr<DevVAddr>,
addrs_count: usize,
out: UserOutPtr<HandleValue>
) -> ZxResult

Pin pages and grant devices access to them.

pub fn sys_pmt_unpin(&self, pmt: HandleValue) -> ZxResult

Unpins pages that were previously pinned by zx_bti_pin().

pub fn sys_bti_release_quarantine(&self, bti: HandleValue) -> ZxResult

Releases all quarantined PMTs for the given BTI.

pub fn sys_pc_firmware_tables(
&self,
resource: HandleValue,
acpi_rsdp_ptr: UserOutPtr<u64>,
smbios_ptr: UserOutPtr<u64>
) -> ZxResult

pub fn sys_interrupt_create(
&self,
resource: HandleValue,
src_num: usize,
options: u32,
out: UserOutPtr<HandleValue>
) -> ZxResult

Creates an interrupt object which represents a physical or virtual interrupt.

pub fn sys_interrupt_bind(
&self,
interrupt: HandleValue,
port: HandleValue,
key: u64,
options: u32
) -> ZxResult

Binds or unbinds an interrupt object to a port.

The key used when binding the interrupt will be present in the key field of the zx_port_packet_t.

pub fn sys_interrupt_trigger(
&self,
interrupt: HandleValue,
options: u32,
timestamp: i64
) -> ZxResult

Triggers a virtual interrupt object.

pub fn sys_interrupt_ack(&self, interrupt: HandleValue) -> ZxResult

Acknowledge an interrupt and re-arm it.

This system call acknowledges an interrupt object, causing it to be eligible to trigger again (and delivering a packet to the port it is bound to).

pub fn sys_interrupt_destroy(&self, interrupt: HandleValue) -> ZxResult

Destroys an interrupt object.

pub async fn sys_interrupt_wait(
&self,
interrupt: HandleValue,
out: UserOutPtr<i64>
) -> ZxResult

A blocking syscall which causes the caller to wait until an interrupt is triggered.

impl Syscall<'_>

pub fn sys_debug_write(&self, buf: UserInPtr<u8>, len: usize) -> ZxResult

Write debug info to the serial port.

pub async fn sys_debug_read(
&self,
handle: HandleValue,
buf: UserOutPtr<u8>,
buf_size: u32,
actual: UserOutPtr<u32>
) -> ZxResult

Read debug info from the serial port.

impl Syscall<'_>

pub fn sys_debuglog_create(
&self,
rsrc: HandleValue,
options: u32,
target: UserOutPtr<HandleValue>
) -> ZxResult

Create a kernel managed debuglog reader or writer.

pub fn sys_debuglog_write(
&self,
handle_value: HandleValue,
options: u32,
buf: UserInPtr<u8>,
len: usize
) -> ZxResult

Write log entry to debuglog.

pub fn sys_debuglog_read(
&self,
handle_value: HandleValue,
options: u32,
buf: UserOutPtr<u8>,
len: usize
) -> ZxResult

Read log entries from debuglog.

impl Syscall<'_>

pub fn sys_create_exception_channel(
&self,
task: HandleValue,
option: u32,
out: UserOutPtr<HandleValue>
) -> ZxResult

Creates a channel which will receive exceptions from the thread, process, or job.

pub fn sys_exception_get_thread(
&self,
exception: HandleValue,
out: UserOutPtr<HandleValue>
) -> ZxResult

Create a handle for the exception’s thread.

The exception handle out will be filled with a new handle to the exception thread.

pub fn sys_exception_get_process(
&self,
exception: HandleValue,
out: UserOutPtr<HandleValue>
) -> ZxResult

Create a handle for the exception’s process.

The exception handle out will be filled with a new handle to the exception process.

Only available for job and process exception channels.
Thread exceptions cannot access their parent process handles.

impl Syscall<'_>

pub fn sys_fifo_create(
&self,
elem_count: usize,
elem_size: usize,
options: u32,
out0: UserOutPtr<HandleValue>,
out1: UserOutPtr<HandleValue>
) -> ZxResult

Creates a fifo, which is actually a pair of fifos of elem_count entries of elem_size bytes.

pub fn sys_fifo_write(
&self,
handle_value: HandleValue,
elem_size: usize,
user_bytes: UserInPtr<u8>,
count: usize,
actual_count_ptr: UserOutPtr<usize>
) -> ZxResult

Write data to a fifo.

pub fn sys_fifo_read(
&self,
handle_value: HandleValue,
elem_size: usize,
user_bytes: UserOutPtr<u8>,
count: usize,
actual_count_ptr: UserOutPtr<usize>
) -> ZxResult

Read data from a fifo.

impl Syscall<'_>

pub async fn sys_futex_wait(
&self,
value_ptr: UserInPtr<AtomicI32>,
current_value: i32,
new_futex_owner: HandleValue,
deadline: Deadline
) -> ZxResult

Wait on a futex.

This system call function atomically verifies that value_ptr still contains the value current_value and sleeps until the futex is made available by a call to zx_futex_wake

pub fn sys_futex_requeue(
&self,
value_ptr: UserInPtr<AtomicI32>,
wake_count: u32,
current_value: i32,
requeue_ptr: UserInPtr<AtomicI32>,
requeue_count: u32,
new_requeue_owner: HandleValue
) -> ZxResult

Wake some waiters and requeue other waiters.

Wake some number of threads waiting on a futex, and move more waiters to another wait queue.

pub fn sys_futex_wake(
&self,
value_ptr: UserInPtr<AtomicI32>,
count: u32
) -> ZxResult

Wake some number of threads waiting on a futex.

Waking up zero threads is not an error condition. Passing in an unallocated address for value_ptr is not an error condition.

pub fn sys_futex_wake_single_owner(
&self,
value_ptr: UserInPtr<AtomicI32>
) -> ZxResult

Wake some number of threads waiting on a futex, and move more waiters to another wait queue.

impl Syscall<'_>

pub fn sys_handle_duplicate(
&self,
handle_value: HandleValue,
rights: u32,
new_handle_value: UserOutPtr<HandleValue>
) -> ZxResult

Creates a duplicate of handle.

Referring to the same underlying object, with new access rights rights.

pub fn sys_handle_close(&self, handle: HandleValue) -> ZxResult

Close a handle and reclaim the underlying object if no other handles to it exist.

pub fn sys_handle_close_many(
&self,
handles: UserInPtr<HandleValue>,
num_handles: usize
) -> ZxResult

Close a number of handles.

pub fn sys_handle_replace(
&self,
handle_value: HandleValue,
rights: u32,
out: UserOutPtr<HandleValue>
) -> ZxResult

Creates a replacement for handle.

Referring to the same underlying object, with new access rights rights.

impl Syscall<'_>

pub fn sys_object_get_property(
&self,
handle_value: HandleValue,
property: u32,
buffer: usize,
buffer_size: usize
) -> ZxResult

Ask for various properties of various kernel objects.

handle_value: HandleValue, indicates the target kernel object. property: u32, indicates which property to get/set. buffer: usize, holds the property value, and must be a pointer to a buffer of value_size bytes.

pub fn sys_object_set_property(
&mut self,
handle_value: HandleValue,
property: u32,
buffer: usize,
buffer_size: usize
) -> ZxResult

Set various properties of various kernel objects.

pub async fn sys_object_wait_one(
&self,
handle: HandleValue,
signals: u32,
deadline: Deadline,
observed: UserOutPtr<Signal>
) -> ZxResult

A blocking syscall waits for signals on an object.

pub fn sys_object_get_info(
&self,
handle: HandleValue,
topic: u32,
buffer: usize,
buffer_size: usize,
actual: UserOutPtr<usize>,
avail: UserOutPtr<usize>
) -> ZxResult

Query information about an object.

topic: u32, indicates what specific information is desired. buffer: usize, a pointer to a buffer of size buffer_size to return the information.

pub fn sys_object_signal_peer(
&self,
handle_value: HandleValue,
clear_mask: u32,
set_mask: u32
) -> ZxResult

Asserts and deasserts the userspace-accessible signal bits on the object’s peer.

pub fn sys_object_wait_async(
&self,
handle_value: HandleValue,
port_handle_value: HandleValue,
key: u64,
signals: u32,
options: u32
) -> ZxResult

A non-blocking syscall subscribes for signals on an object.

pub fn sys_object_signal(
&self,
handle_value: HandleValue,
clear_mask: u32,
set_mask: u32
) -> ZxResult

Signal an object.

Asserts and deasserts the userspace-accessible signal bits on an object.

pub async fn sys_object_wait_many(
&self,
user_items: UserInOutPtr<UserWaitItem>,
count: u32,
deadline: Deadline
) -> ZxResult

Wait for signals on multiple objects.

pub fn sys_object_get_child(
&self,
handle: HandleValue,
koid: KoID,
rights: u32,
out: UserOutPtr<HandleValue>
) -> ZxResult

Find the child of an object by its kid.

Given a kernel object with children objects, obtain a handle to the child specified by the provided kernel object id.

impl Syscall<'_>

pub fn sys_pci_add_subtract_io_range(
&self,
handle: HandleValue,
mmio: bool,
base: u64,
len: u64,
add: bool
) -> ZxResult

pub fn sys_pci_cfg_pio_rw(
&self,
handle: HandleValue,
bus: u8,
dev: u8,
func: u8,
offset: u8,
value_ptr: UserInOutPtr<u32>,
width: usize,
write: bool
) -> ZxResult

pub fn sys_pci_init(
&self,
handle: HandleValue,
init_buf: usize,
len: u32
) -> ZxResult

pub fn sys_pci_map_interrupt(
&self,
dev: HandleValue,
irq: i32,
out_handle: UserOutPtr<HandleValue>
) -> ZxResult

pub fn sys_pci_get_nth_device(
&self,
handle: HandleValue,
index: u32,
out_info: UserOutPtr<PcieDeviceInfo>,
out_handle: UserOutPtr<HandleValue>
) -> ZxResult

pub fn sys_pci_get_bar(
&self,
handle: HandleValue,
bar_num: u32,
out_bar: UserOutPtr<PciBar>,
out_handle: UserOutPtr<HandleValue>
) -> ZxResult

pub fn sys_pci_enable_bus_master(
&self,
handle: HandleValue,
enable: bool
) -> ZxResult

pub fn sys_pci_query_irq_mode(
&self,
handle: HandleValue,
mode: u32,
out_max_irqs: UserOutPtr<u32>
) -> ZxResult

pub fn sys_pci_set_irq_mode(
&self,
handle: HandleValue,
mode: u32,
requested_irq_count: u32
) -> ZxResult

pub fn sys_pci_config_read(
&self,
handle: HandleValue,
offset: usize,
width: usize,
out_val: UserOutPtr<u32>
) -> ZxResult

pub fn sys_pci_config_write(
&self,
handle: HandleValue,
offset: usize,
width: usize,
value: u32
) -> ZxResult

impl Syscall<'_>

pub fn sys_port_create(
&self,
options: u32,
out: UserOutPtr<HandleValue>
) -> ZxResult

Create an IO port.

pub async fn sys_port_wait(
&self,
handle_value: HandleValue,
deadline: Deadline,
packet_res: UserOutPtr<PortPacket>
) -> ZxResult

Wait for a packet arrival in a port.

pub fn sys_port_queue(
&self,
handle_value: HandleValue,
packcet_in: UserInPtr<PortPacket>
) -> ZxResult

Queue a packet to a port.

impl Syscall<'_>

pub fn sys_resource_create(
&self,
parent_rsrc: HandleValue,
options: u32,
base: u64,
size: u64,
name: UserInPtr<u8>,
name_size: u64,
out: UserOutPtr<HandleValue>
) -> ZxResult

Create a resource object for use with other DDK syscalls.

impl Syscall<'_>

pub fn sys_timer_create(
&self,
options: u32,
clock_id: u32,
out: UserOutPtr<HandleValue>
) -> ZxResult

Create a timer.

The timer is an object that can signal when a specified point in time has been reached.

pub fn sys_event_create(
&self,
options: u32,
out: UserOutPtr<HandleValue>
) -> ZxResult

Create an event.

pub fn sys_eventpair_create(
&self,
options: u32,
out0: UserOutPtr<HandleValue>,
out1: UserOutPtr<HandleValue>
) -> ZxResult

Create an event pair.

pub fn sys_timer_set(
&self,
handle: HandleValue,
deadline: Deadline,
slack: i64
) -> ZxResult

Start a timer.

To fire the timer immediately pass a deadline less than or equal to 0.
The slack parameter specifies a range from deadline - slack to deadline + slack during which the timer is allowed to fire.

pub fn sys_timer_cancel(&self, handle: HandleValue) -> ZxResult

Cancel a timer.

impl Syscall<'_>

pub fn sys_socket_create(
&self,
options: u32,
out0: UserOutPtr<HandleValue>,
out1: UserOutPtr<HandleValue>
) -> ZxResult

Create a socket.

Socket is a connected pair of bidirectional stream transports, that can move only data, and that have a maximum capacity.

pub fn sys_socket_write(
&self,
handle_value: HandleValue,
options: u32,
user_bytes: UserInPtr<u8>,
count: usize,
actual_count_ptr: UserOutPtr<usize>
) -> ZxResult

Write data to a socket.

Attempts to write count: usize bytes to the socket specified by handle_value.

pub fn sys_socket_read(
&self,
handle_value: HandleValue,
options: u32,
user_bytes: UserOutPtr<u8>,
count: usize,
actual_count_ptr: UserOutPtr<usize>
) -> ZxResult

Read data from a socket.

pub fn sys_socket_shutdown(&self, socket: HandleValue, options: u32) -> ZxResult

Prevent future reading or writing on a socket.

impl Syscall<'_>

pub fn sys_stream_create(
&self,
options: u32,
vmo_handle: HandleValue,
seek: usize,
out: UserOutPtr<HandleValue>
) -> ZxResult

Create a stream from a VMO.

Stream for reads and writes the data in an underlying VMO.

pub fn sys_stream_writev(
&self,
handle_value: HandleValue,
options: u32,
vector: UserInPtr<IoVecIn>,
vector_size: usize,
actual_count_ptr: UserOutPtr<usize>
) -> ZxResult

Write data to a stream at the current seek offset.

pub fn sys_stream_writev_at(
&self,
handle_value: HandleValue,
options: u32,
offset: usize,
vector: UserInPtr<IoVecIn>,
vector_size: usize,
actual_count_ptr: UserOutPtr<usize>
) -> ZxResult

Write data to a stream at the given offset.

pub fn sys_stream_readv(
&self,
handle_value: HandleValue,
options: u32,
vector: UserInPtr<IoVecOut>,
vector_size: usize,
actual_count_ptr: UserOutPtr<usize>
) -> ZxResult

Read data from a stream at the current seek offset.

pub fn sys_stream_readv_at(
&self,
handle_value: HandleValue,
options: u32,
offset: usize,
vector: UserInPtr<IoVecOut>,
vector_size: usize,
actual_count_ptr: UserOutPtr<usize>
) -> ZxResult

Read data from a stream at the given offset.

pub fn sys_stream_seek(
&self,
handle_value: HandleValue,
whence: usize,
offset: isize,
out_seek: UserOutPtr<usize>
) -> ZxResult

Modify the seek offset.

Sets the seek offset of the stream to offset relative to whence.

impl Syscall<'_>

pub fn sys_system_get_event(
&self,
root_job: HandleValue,
kind: u32,
out: UserOutPtr<HandleValue>
) -> ZxResult

Retrieve a handle to a system event.

root_job: HandleValue, must be a handle to the root job of the system.
kind: u32, must be one of the following:

    const EVENT_OUT_OF_MEMORY: u32 = 1;
    const EVENT_MEMORY_PRESSURE_CRITICAL: u32 = 2;
    const EVENT_MEMORY_PRESSURE_WARNING: u32 = 3;
    const EVENT_MEMORY_PRESSURE_NORMAL: u32 = 4;

impl Syscall<'_>

pub fn sys_process_create(
&self,
job: HandleValue,
name: UserInPtr<u8>,
name_size: usize,
options: u32,
proc_handle: UserOutPtr<HandleValue>,
vmar_handle: UserOutPtr<HandleValue>
) -> ZxResult

Create a new process.

Upon success, handles for the new process and the root of its address space are returned.

pub fn sys_process_exit(&mut self, code: i64) -> ZxResult

Exits the currently running process.

pub fn sys_thread_create(
&self,
proc_handle: HandleValue,
name: UserInPtr<u8>,
name_size: usize,
options: u32,
thread_handle: UserOutPtr<HandleValue>
) -> ZxResult

Creates a thread within the specified process.

Upon success a handle for the new thread is returned.

pub fn sys_process_start(
&self,
proc_handle: HandleValue,
thread_handle: HandleValue,
entry: usize,
stack: usize,
arg1_handle: HandleValue,
arg2: usize
) -> ZxResult

Start execution on a process.

This system call is similar to zx_thread_start(), but is used for the purpose of starting the first thread in a process.

pub fn sys_thread_read_state(
&self,
handle: HandleValue,
kind: u32,
buffer: UserOutPtr<u8>,
buffer_size: usize
) -> ZxResult

Read one aspect of thread state.

The thread state may only be written when the thread is halted for an exception or the thread is suspended.

pub fn sys_thread_write_state(
&self,
handle: HandleValue,
kind: u32,
buffer: UserInPtr<u8>,
buffer_size: usize
) -> ZxResult

Write one aspect of thread state.

The thread state may only be written when the thread is halted for an exception or the thread is suspended.

pub fn sys_job_set_critical(
&self,
job_handle: HandleValue,
options: u32,
process_handle: HandleValue
) -> ZxResult

Sets process as critical to job.

When process terminates, job will be terminated as if zx_task_kill() was called on it.

pub fn sys_thread_start(
&self,
handle_value: HandleValue,
entry: usize,
stack: usize,
arg1: usize,
arg2: usize
) -> ZxResult

Start execution on a thread.

pub fn sys_thread_exit(&mut self) -> ZxResult

Terminate the current running thread.

Causes the currently running thread to cease running and exit.

pub fn sys_task_suspend_token(
&self,
handle: HandleValue,
token: UserOutPtr<HandleValue>
) -> ZxResult

Suspend the given task.

This function replaces task_suspend. When all callers are updated, zx_task_suspend() will be deleted and this function will be renamed zx_task_suspend().

pub fn sys_task_kill(&mut self, handle: HandleValue) -> ZxResult

Kill the provided task (job, process, or thread).

pub fn sys_job_create(
&self,
parent: HandleValue,
options: u32,
out: UserOutPtr<HandleValue>
) -> ZxResult

Create a new child job object given a parent job.

pub fn sys_job_set_policy(
&self,
handle: HandleValue,
options: u32,
topic: u32,
policy: usize,
count: u32
) -> ZxResult

Sets one or more security and/or resource policies to an empty job.

pub fn sys_process_read_memory(
&self,
handle_value: HandleValue,
vaddr: usize,
buffer: UserOutPtr<u8>,
buffer_size: usize,
actual: UserOutPtr<usize>
) -> ZxResult

Read from the given process’s address space.

This function will eventually be replaced with something vmo-centric.

pub fn sys_process_write_memory(
&self,
handle_value: HandleValue,
vaddr: usize,
buffer: UserInPtr<u8>,
buffer_size: usize,
actual: UserOutPtr<usize>
) -> ZxResult

Write into the given process’s address space.

impl Syscall<'_>

pub fn sys_clock_create(
&self,
_options: u64,
_user_args: UserInPtr<u8>,
_out: UserOutPtr<HandleValue>
) -> ZxResult

Create a new clock object.

pub fn sys_clock_get(&self, clock_id: u32, time: UserOutPtr<u64>) -> ZxResult

Acquire the current time.

Returns the current time of clock_id via time.

Returns whether clock_id was valid.

pub fn sys_clock_read(
&self,
handle: HandleValue,
now: UserOutPtr<u64>
) -> ZxResult

Perform a basic read of the clock.

pub fn sys_clock_adjust(
&self,
resource: HandleValue,
clock_id: u32,
offset: u64
) -> ZxResult

pub fn sys_clock_update(
&self,
_handle: HandleValue,
_options: u64,
_user_args: UserInPtr<u8>
) -> ZxResult

Make adjustments to a clock object.

pub async fn sys_nanosleep(&self, deadline: Deadline) -> ZxResult

Sleep for some number of nanoseconds.

A deadline value less than or equal to 0 immediately yields the thread.

impl Syscall<'_>

pub fn sys_vmar_allocate(
&self,
parent_vmar: HandleValue,
options: u32,
offset: u64,
size: u64,
out_child_vmar: UserOutPtr<HandleValue>,
out_child_addr: UserOutPtr<usize>
) -> ZxResult

Allocate a new subregion.

Creates a new VMAR within the one specified by parent_vmar.

pub fn sys_vmar_map(
&self,
vmar_handle: HandleValue,
options: u32,
vmar_offset: usize,
vmo_handle: HandleValue,
vmo_offset: usize,
len: usize,
mapped_addr: UserOutPtr<VirtAddr>
) -> ZxResult

Add a memory mapping.

Maps the given VMO into the given virtual memory address region.

pub fn sys_vmar_destroy(&self, handle_value: HandleValue) -> ZxResult

Destroy a virtual memory address region.

Unmaps all mappings within the given region, and destroys all sub-regions of the region.

This operation is logically recursive.

pub fn sys_vmar_protect(
&self,
handle_value: HandleValue,
options: u32,
addr: u64,
len: u64
) -> ZxResult

Set protection of virtual memory pages.

pub fn sys_vmar_unmap(
&self,
handle_value: HandleValue,
addr: usize,
len: usize
) -> ZxResult

Unmap virtual memory pages.

impl Syscall<'_>

pub fn sys_vmo_create(
&self,
size: u64,
options: u32,
out: UserOutPtr<HandleValue>
) -> ZxResult

Create a new virtual memory object(VMO).

pub fn sys_vmo_read(
&self,
handle_value: HandleValue,
buf: UserOutPtr<u8>,
offset: u64,
buf_size: usize
) -> ZxResult

Read bytes from a VMO.

pub fn sys_vmo_write(
&self,
handle_value: HandleValue,
buf: UserInPtr<u8>,
offset: u64,
buf_size: usize
) -> ZxResult

Write bytes to a VMO.

pub fn sys_vmo_replace_as_executable(
&self,
handle: HandleValue,
vmex: HandleValue,
out: UserOutPtr<HandleValue>
) -> ZxResult

Add execute rights to a VMO.

pub fn sys_vmo_get_size(
&self,
handle: HandleValue,
size: UserOutPtr<usize>
) -> ZxResult

Obtain the current size of a VMO object.

pub fn sys_vmo_create_child(
&self,
handle_value: HandleValue,
options: u32,
offset: usize,
size: usize,
out: UserOutPtr<HandleValue>
) -> ZxResult

Create a child of an existing VMO (new virtual memory object).

pub fn sys_vmo_create_physical(
&self,
resource: HandleValue,
paddr: PhysAddr,
size: usize,
out: UserOutPtr<HandleValue>
) -> ZxResult

Create a VM object referring to a specific contiguous range of physical memory.

pub fn sys_vmo_create_contiguous(
&self,
bti: HandleValue,
size: usize,
align_log2: u32,
out: UserOutPtr<HandleValue>
) -> ZxResult

Create a VM object referring to a specific contiguous range of physical frame.

pub fn sys_vmo_set_size(
&self,
handle_value: HandleValue,
size: usize
) -> ZxResult

Resize a VMO object.

pub fn sys_vmo_op_range(
&self,
handle_value: HandleValue,
op: u32,
offset: usize,
len: usize,
_buffer: UserOutPtr<u8>,
_buffer_size: usize
) -> ZxResult

Perform an operation on a range of a VMO.

Performs cache and memory operations against pages held by the VMO.

pub fn sys_vmo_cache_policy(
&self,
handle_value: HandleValue,
policy: u32
) -> ZxResult

Set the caching policy for pages held by a VMO.

impl Syscall<'_>

pub async fn syscall(&mut self, num: u32, args: [usize; 8]) -> isize